requirements.txt
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.6.0
pymongo==4.3.3
Werkzeug==2.2.3
